+++ lib/sidekiq_publisher/worker.rb
@@ -3,7 +3,7 @@
 module SidekiqPublisher
   module Worker
     def self.included(base)
-      base.include(Sidekiq::Worker)
+      base.include(SidekiqPublisher::Compatibility.sidekiq_job_class)
       base.singleton_class.public_send(:alias_method, :sidekiq_client_push, :client_push)
       base.extend(ClassMethods)
     end